refactor: move routes to APIRouter
